refactor(cognition): extract responses oauth providers
