+++ loom/core/cognition/providers.py
@@ -44,172 +44,6 @@ from .forensics import get_forensics
 logger = logging.getLogger(__name__)
 
 
-def _stream_interrupt_detail(provider: str, exc: BaseException) -> str:
-    """Format a streaming-transport error.
-
-    Some httpx exceptions (notably ``RemoteProtocolError`` raised when an
-    SSE peer closes mid-stream) carry an empty ``str(exc)``. Fall back
-    to the exception class name so the user never sees the bare
-    ``turn aborted with error: `` with nothing after the colon.
-    """
-    detail = str(exc).strip() or type(exc).__name__
-    return f"{provider} stream interrupted: {detail}"
-
-
-async def _http_status_detail(provider: str, exc: Any) -> str:
-    """Format an HTTPStatusError into a human-readable detail string.
-
-    Works for both regular and streaming responses. Streaming responses
-    raised through ``client.stream(...).__aenter__()`` haven't loaded
-    their body yet, so accessing ``.text`` would raise
-    ``httpx.ResponseNotRead``. We call ``aread()`` first and swallow any
-    read failure so the caller always gets *some* detail.
-    """
-    response = getattr(exc, "response", None)
-    status_code = getattr(response, "status_code", "unknown")
-    body = ""
-    if response is not None:
-        aread = getattr(response, "aread", None)
-        if callable(aread):
-            try:
-                await aread()
-            except Exception:
-                pass
-        try:
-            body = str(getattr(response, "text", "") or "").strip()
-        except Exception:
-            body = ""
-    detail = f"{provider} request failed with HTTP {status_code}"
-    if body:
-        detail += f": {body[:1000]}"
-    return detail
-
-
-_RESPONSES_REASONING_EFFORTS = ("none", "minimal", "low", "medium", "high", "xhigh")
-_RESPONSES_REASONING_DEFAULT_EFFORT = "high"
-# Per-model accepted values are not uniform: gpt-5.5 supports
-# none/low/medium/high/xhigh but rejects "minimal"; xAI grok-4.x accepts
-# all six. We pass any value in this set through to the backend; a per-model
-# 400 is surfaced via PR #453's SSE failure handling so the user gets a
-# clear "Unsupported value 'X' for model Y" message instead of a silent
-# fallback that drifts from what they configured.
-
-
-class _ReasoningStreamWrapper:
-    """Wrap Responses SSE reasoning_summary deltas as ``<think>…</think>`` text.
-
-    The session-level consumer (``session.py``) already parses ``<think>``
-    tags out of the assistant text stream and surfaces them as
-    ``ThinkCollapsed`` UI events. Wrapping reasoning chunks in the same
-    protocol lets the user see *that the model is thinking* during the long
-    reasoning phase of o1-class / gpt-5.5 / grok-4 reasoning models —
-    instead of staring at an empty stream until output_text starts.
-
-    State: ``False`` until first reasoning delta seen; flipped back when
-    reasoning is done OR output_text starts OR stream ends. The caller is
-    responsible for calling ``flush()`` after the SSE loop exits to close
-    any dangling tag.
-    """
-
-    __slots__ = ("_open",)
-
-    def __init__(self) -> None:
-        self._open = False
-
-    def open_chunk(self, delta: str) -> str:
-        if not self._open:
-            self._open = True
-            return f"<think>{delta}"
-        return delta
-
-    def close(self) -> str:
-        if self._open:
-            self._open = False
-            return "</think>"
-        return ""
-
-    @property
-    def is_open(self) -> bool:
-        return self._open
-
-
-def _normalize_reasoning_effort(value: str | None) -> str:
-    """Coerce an optional config value into a valid Responses reasoning effort.
-
-    Falls back to the default for empty/None/invalid input rather than
-    raising: a typo in loom.toml should not block the whole chat path.
-    """
-    if not value:
-        return _RESPONSES_REASONING_DEFAULT_EFFORT
-    candidate = str(value).strip().lower()
-    if candidate in _RESPONSES_REASONING_EFFORTS:
-        return candidate
-    return _RESPONSES_REASONING_DEFAULT_EFFORT
-
-
-def _responses_sse_error_detail(provider: str, event_type: str, data: dict[str, Any]) -> str | None:
-    """Return a user-facing error detail for Responses SSE error events.
-
-    Some Responses-compatible backends report request failures inside the
-    event stream after HTTP 200. If we ignore those events, the harness sees
-    an empty successful turn and the operator gets no actionable error.
-    """
-    response = data.get("response")
-    response = response if isinstance(response, dict) else {}
-    status = str(response.get("status") or "").lower()
-    error = data.get("error") or response.get("error")
-    incomplete = response.get("incomplete_details")
-
-    if status == "incomplete" and isinstance(incomplete, dict):
-        reason = str(incomplete.get("reason") or "").strip()
-        # Preserve existing max-output handling: session.stream_turn can
-        # recover from max_tokens by injecting its reasoning-continuation
-        # reminder. Other incomplete reasons need to be visible.
-        if reason and reason != "max_output_tokens":
-            error = incomplete
-
-    failed_event = (
-        event_type in {"error", "response.failed"}
-        or event_type.endswith(".failed")
-        or status in {"failed", "cancelled"}
-        or error is not None
-    )
-    if not failed_event:
-        return None
-
-    parts: list[str] = []
-    if event_type:
-        parts.append(event_type)
-    if status:
-        parts.append(f"status={status}")
-
-    detail = ""
-    if isinstance(error, dict):
-        message = error.get("message") or error.get("detail") or error.get("reason")
-        code = error.get("code") or error.get("type")
-        if code and message:
-            detail = f"{code}: {message}"
-        elif message:
-            detail = str(message)
-        elif code:
-            detail = str(code)
-        else:
-            detail = json.dumps(error, ensure_ascii=False)
-    elif error:
-        detail = str(error)
-    elif isinstance(incomplete, dict):
-        detail = json.dumps(incomplete, ensure_ascii=False)
-    elif data.get("message"):
-        detail = str(data.get("message"))
-
-    context = " ".join(parts).strip()
-    if detail:
-        return f"{provider} response error ({context}): {detail[:1000]}"
-    if context:
-        return f"{provider} response error ({context})"
-    return f"{provider} response error"
-
-
 # ---------------------------------------------------------------------------
 # Retry helper
 # ---------------------------------------------------------------------------
@@ -1048,632 +882,9 @@ class OpenAIProvider(_OpenAICompatibleBase):
     MAX_TOKENS_PARAM = "max_completion_tokens"
 
 
-class CodexResponsesProvider(LLMProvider):
-    """
-    OpenAI Codex OAuth chat via ChatGPT's Codex Responses backend.
 
-    Routing prefix: ``codex/``. This provider intentionally does not fall back
-    to ``OPENAI_API_KEY`` so users do not accidentally burn API quota when they
-    selected Codex OAuth mode.
-    """
-
-    name = "codex"
-    ROUTING_PREFIX = "codex/"
-    DEFAULT_MODEL = "gpt-5.5"
-    DEFAULT_BASE_URL = "https://chatgpt.com/backend-api/codex/responses"
-    # gpt-5.5 at ``effort=high`` on big multi-turn payloads routinely
-    # spends 3–6 minutes thinking before the first output token. The old
-    # 180s default raised ``httpx.ReadTimeout`` mid-reasoning and the
-    # harness surfaced ``turn aborted with error: ... ReadTimeout`` — DK
-    # observed this 3 times in a single PR review session. 600s gives
-    # enough headroom; users can override via ``[providers.codex].timeout``.
-    DEFAULT_TIMEOUT = 600.0
-    SUPPORTS_MAX_OUTPUT_TOKENS = False
-
-    def __init__(
-        self,
-        model: str = "",
-        base_url: str = "",
-        timeout: float = 0.0,
-        reasoning_effort: str | None = None,
-    ) -> None:
-        self.model = model or (self.ROUTING_PREFIX + self.DEFAULT_MODEL)
-        self._base_url = base_url or self.DEFAULT_BASE_URL
-        self._timeout = timeout or self.DEFAULT_TIMEOUT
-        # Always send a ``reasoning.effort`` value. Omitting the field lets
-        # the ChatGPT.com Codex backend run gpt-5.5's reasoning *unbounded*
-        # — in large-context multi-turn (real Loom chat with skills + tool
-        # history) that means 3+ minute silent stalls before the first
-        # output token. Pre-PR #449 this was implicitly capped by
-        # ``max_output_tokens`` (which the backend now rejects), so the cap
-        # has to move to ``reasoning.effort`` instead. Default ``high``:
-        # tier 2 users selected codex/gpt-5.5 because they want depth, and
-        # all effort levels finish in <6s on big multi-turn — only the
-        # *unbounded* path stalls. ``summary`` is intentionally never sent:
-        # PR #455 showed ``summary: auto`` is what triples TTFT, not the
-        # effort field itself.
-        self._reasoning_effort = _normalize_reasoning_effort(reasoning_effort)
-
-    def _api_model(self) -> str:
-        return self.model.removeprefix(self.ROUTING_PREFIX)
-
-    def _load_bearer_token(self) -> str:
-        from loom.core.cognition.openai_auth import load_codex_oauth_credential
-
-        credential = load_codex_oauth_credential()
-        if credential is None:
-            raise RuntimeError(
-                "No unexpired Codex OAuth token found. Run `codex login` "
-                "or switch to an API-key OpenAI model such as `gpt-5.5`."
-            )
-        return credential.token
-
-    def _build_payload(
-        self,
-        messages: list[dict[str, Any]],
-        tools: list[dict[str, Any]] | None,
-        max_tokens: int,
-    ) -> dict[str, Any]:
-        instructions: list[str] = []
-        input_items: list[dict[str, Any]] = []
-        for msg in messages:
-            role = msg.get("role")
-            content = msg.get("content", "")
-            if role == "system":
-                if content:
-                    instructions.append(str(content))
-                continue
-            if role == "tool":
-                call_id = msg.get("tool_call_id")
-                if call_id:
-                    input_items.append({
-                        "type": "function_call_output",
-                        "call_id": call_id,
-                        "output": str(content),
-                    })
-                continue
-            if role == "assistant":
-                if content:
-                    input_items.append({
-                        "role": "assistant",
-                        "content": [{"type": "output_text", "text": str(content)}],
-                    })
-                for tc in msg.get("tool_calls", []) or []:
-                    fn = tc.get("function", {}) if isinstance(tc, dict) else {}
-                    input_items.append({
-                        "type": "function_call",
-                        "call_id": tc.get("id") or str(uuid.uuid4()),
-                        "name": fn.get("name", ""),
-                        "arguments": fn.get("arguments", "{}"),
-                    })
-                continue
-            input_items.append({
-                "role": "user" if role != "assistant" else "assistant",
-                "content": [{"type": "input_text", "text": str(content)}],
-            })
-
-        payload: dict[str, Any] = {
-            "model": self._api_model(),
-            "instructions": "\n\n".join(instructions) or "You are a helpful assistant.",
-            "input": input_items,
-            "stream": True,
-            "store": False,
-            # See ``__init__`` for the full story. ``effort`` is the
-            # required reasoning-budget signal; ``summary`` is omitted on
-            # purpose — emitting it requires the backend to allocate a
-            # parallel reasoning_summary stream that triples TTFT.
-            "reasoning": {"effort": self._reasoning_effort},
-        }
-        if self.SUPPORTS_MAX_OUTPUT_TOKENS:
-            payload["max_output_tokens"] = max_tokens
-        if tools:
-            payload["tools"] = self.format_tools(tools)
-        return payload
-
-    def format_tools(self, tools: list[dict[str, Any]]) -> list[dict[str, Any]]:
-        formatted: list[dict[str, Any]] = []
-        for tool in tools:
-            formatted.append({
-                "type": "function",
-                "name": tool.get("name", ""),
-                "description": tool.get("description", ""),
-                "parameters": tool.get("parameters") or tool.get("input_schema") or {},
-            })
-        return formatted
-
-    async def chat(
-        self,
-        messages: list[dict[str, Any]],
-        tools: list[dict[str, Any]] | None = None,
-        max_tokens: int = 8096,
-    ) -> LLMResponse:
-        text_parts: list[str] = []
-        final: LLMResponse | None = None
-        async for chunk, resp in self.stream_chat(
-            messages=messages,
-            tools=tools,
-            max_tokens=max_tokens,
-        ):
-            if resp is not None:
-                final = resp
-            elif chunk:
-                text_parts.append(chunk)
-        return final or LLMResponse(
-            text="".join(text_parts) or None,
-            tool_uses=[],
-            stop_reason="end_turn",
-        )
-
-    async def stream_chat(
-        self,
-        messages: list[dict[str, Any]],
-        tools: list[dict[str, Any]] | None = None,
-        max_tokens: int = 8096,
-        *,
-        abort_signal: Any = None,
-    ) -> AsyncIterator[tuple[str, LLMResponse | None]]:
-        import httpx
-
-        bearer = self._load_bearer_token()
-        payload = self._build_payload(messages, tools, max_tokens)
-        full_content = ""
-        output_items: list[dict[str, Any]] = []
-        input_tokens = 0
-        output_tokens = 0
-        response_status = "in_progress"
-        reasoning_wrapper = _ReasoningStreamWrapper()
-
-        # SSE may already have yielded partial output; retrying mid-stream can
-        # duplicate content or tool calls, so failures propagate to the caller.
-        async with httpx.AsyncClient(timeout=self._timeout) as client:
-            async with client.stream(
-                "POST",
-                self._base_url,
-                headers={
-                    "Authorization": f"Bearer {bearer}",
-                    "Content-Type": "application/json",
-                    "Accept": "text/event-stream",
-                },
-                json=payload,
-            ) as resp:
-                try:
-                    resp.raise_for_status()
-                except httpx.HTTPStatusError as exc:
-                    raise RuntimeError(await _http_status_detail("Codex Responses", exc)) from exc
-                event: str | None = None
-                data_lines: list[str] = []
-                try:
-                    async for line in resp.aiter_lines():
-                        if abort_signal is not None and abort_signal.is_set():
-                            break
-                        if line.startswith("event: "):
-                            event = line[7:]
-                        elif line.startswith("data: "):
-                            data_lines.append(line[6:])
-                        elif line == "" and event:
-                            raw = "\n".join(data_lines)
-                            event = None
-                            data_lines = []
-                            try:
-                                data = json.loads(raw)
-                            except json.JSONDecodeError:
-                                continue
-                            event_type = data.get("type", "")
-                            if detail := _responses_sse_error_detail(
-                                "Codex Responses", event_type, data,
-                            ):
-                                raise RuntimeError(detail)
-                            delta = data.get("delta")
-                            # OpenAI Responses SSE emits ``delta`` on multiple
-                            # event types — only ``response.output_text.delta``
-                            # is user-facing text. ``response.function_call_arguments.delta``
-                            # streams the JSON args of an in-flight tool call
-                            # and must NOT leak into the assistant message;
-                            # the canonical args still arrive on
-                            # ``response.output_item.done`` so we can ignore
-                            # the streaming form entirely.
-                            #
-                            # ``reasoning_summary_text.delta`` carries the
-                            # backend's reasoning summary — surface it wrapped
-                            # in ``<think>…</think>`` so session.stream_turn
-                            # renders it as a collapsed thinking block instead
-                            # of a silent stall.
-                            if isinstance(delta, str) and event_type.endswith("reasoning_summary_text.delta"):
-                                yield (reasoning_wrapper.open_chunk(delta), None)
-                            elif (
-                                event_type.endswith("reasoning_summary_text.done")
-                                or event_type.endswith("reasoning_summary_part.done")
-                            ):
-                                # OpenAI Responses emits BOTH events at the
-                                # end of a summary part. ``part.done`` is the
-                                # outer wrapper and is the one we can rely on
-                                # to always arrive — ``text.done`` may be
-                                # omitted by some compat backends. Either
-                                # event closes the ``<think>`` tag so the UI
-                                # surfaces the block before ``output_text``
-                                # starts, which is the whole point of the fix.
-                                closing = reasoning_wrapper.close()
-                                if closing:
-                                    yield (closing, None)
-                            elif isinstance(delta, str) and event_type.endswith("output_text.delta"):
-                                closing = reasoning_wrapper.close()
-                                if closing:
-                                    yield (closing, None)
-                                full_content += delta
-                                yield (delta, None)
-                            item = data.get("item")
-                            if isinstance(item, dict) and event_type.endswith("output_item.done"):
-                                output_items.append(item)
-                            response = data.get("response")
-                            if isinstance(response, dict):
-                                response_status = str(response.get("status") or response_status)
-                                usage = response.get("usage") or {}
-                                if isinstance(usage, dict):
-                                    input_tokens = int(
-                                        usage.get("input_tokens") or usage.get("prompt_tokens") or input_tokens
-                                    )
-                                    output_tokens = int(
-                                        usage.get("output_tokens") or usage.get("completion_tokens") or output_tokens
-                                    )
-                except httpx.HTTPError as exc:
-                    raise RuntimeError(
-                        _stream_interrupt_detail("Codex Responses", exc)
-                    ) from exc
-
-        # Close any reasoning tag the stream left dangling (e.g. abort signal,
-        # or backend that omitted the .done event).
-        closing = reasoning_wrapper.close()
-        if closing:
-            yield (closing, None)
-
-        tool_uses: list[ToolUse] = []
-        raw_tool_calls: list[dict[str, Any]] = []
-        for item in output_items:
-            if item.get("type") != "function_call":
-                continue
-            args_raw = item.get("arguments") or "{}"
-            try:
-                args = json.loads(args_raw)
-            except (json.JSONDecodeError, ValueError):
-                args = {"_raw": args_raw}
-            call_id = item.get("call_id") or item.get("id") or str(uuid.uuid4())
-            name = item.get("name") or ""
-            tool_uses.append(ToolUse(id=call_id, name=name, args=args))
-            raw_tool_calls.append({
-                "id": call_id,
-                "type": "function",
-                "function": {"name": name, "arguments": args_raw},
-            })
-
-        stop_reason = "tool_use" if tool_uses else "end_turn"
-        if response_status == "incomplete":
-            stop_reason = "max_tokens"
-
-        raw_message: dict[str, Any] = {"role": "assistant", "content": full_content}
-        if raw_tool_calls:
-            raw_message["tool_calls"] = raw_tool_calls
-        yield ("", LLMResponse(
-            text=full_content or None,
-            tool_uses=tool_uses,
-            stop_reason=stop_reason,
-            input_tokens=input_tokens,
-            output_tokens=output_tokens,
-            raw_message=raw_message,
-        ))
-
-    def format_tool_result(
-        self, tool_use_id: str, content: str, success: bool = True
-    ) -> dict[str, Any]:
-        return {
-            "role": "tool",
-            "tool_call_id": tool_use_id,
-            "content": content if success else f"Error: {content}",
-        }
-
-
-class XAIResponsesProvider(LLMProvider):
-    """
-    xAI Grok OAuth chat via xAI's Responses API.
-
-    Routing prefix: ``xai/``. This provider intentionally has no
-    ``XAI_API_KEY`` fallback; selecting ``xai/...`` means OAuth only.
-    """
-
-    name = "xai"
-    ROUTING_PREFIX = "xai/"
-    DEFAULT_MODEL = "grok-4.3"
-    DEFAULT_BASE_URL = "https://api.x.ai/v1/responses"
-    # Match Codex (600s) — grok-4.x at ``effort=high`` on big context can
-    # also stretch past the old 180s ceiling. Users can override via
-    # ``[providers.xai_oauth].timeout``.
-    DEFAULT_TIMEOUT = 600.0
-
-    def __init__(
-        self,
-        model: str = "",
-        base_url: str = "",
-        timeout: float = 0.0,
-        reasoning_effort: str | None = None,
-    ) -> None:
-        self.model = model or (self.ROUTING_PREFIX + self.DEFAULT_MODEL)
-        self._base_url = base_url or self.DEFAULT_BASE_URL
-        self._timeout = timeout or self.DEFAULT_TIMEOUT
-        # Match Codex: default ``high`` since tier-2 model selection
-        # implies the user wants reasoning depth. Unlike Codex, omitting
-        # the field doesn't stall on xAI's backend — but sending it makes
-        # behavior consistent across providers and gives the user a single
-        # knob to turn. xAI accepts the full range including ``minimal``
-        # (which gpt-5.5 rejects).
-        self._reasoning_effort = _normalize_reasoning_effort(reasoning_effort)
-
-    def _api_model(self) -> str:
-        return self.model.removeprefix(self.ROUTING_PREFIX)
-
-    def _load_bearer_token(self) -> str:
-        from loom.core.cognition.xai_auth import load_xai_oauth_credential
-
-        credential = load_xai_oauth_credential()
-        if credential is None:
-            raise RuntimeError(
-                "No unexpired xAI OAuth token found. Run `loom auth xai` "
-                "before using an OAuth model such as `xai/grok-4.3`."
-            )
-        return credential.token
-
-    def _build_payload(
-        self,
-        messages: list[dict[str, Any]],
-        tools: list[dict[str, Any]] | None,
-        max_tokens: int,
-    ) -> dict[str, Any]:
-        instructions: list[str] = []
-        input_items: list[dict[str, Any]] = []
-        for msg in messages:
-            role = msg.get("role")
-            content = msg.get("content", "")
-            if role == "system":
-                if content:
-                    instructions.append(str(content))
-                continue
-            if role == "tool":
-                call_id = msg.get("tool_call_id")
-                if call_id:
-                    input_items.append({
-                        "type": "function_call_output",
-                        "call_id": call_id,
-                        "output": str(content),
-                    })
-                continue
-            if role == "assistant":
-                if content:
-                    input_items.append({
-                        "role": "assistant",
-                        "content": [{"type": "output_text", "text": str(content)}],
-                    })
-                for tc in msg.get("tool_calls", []) or []:
-                    fn = tc.get("function", {}) if isinstance(tc, dict) else {}
-                    input_items.append({
-                        "type": "function_call",
-                        "call_id": tc.get("id") or str(uuid.uuid4()),
-                        "name": fn.get("name", ""),
-                        "arguments": fn.get("arguments", "{}"),
-                    })
-                continue
-            input_items.append({
-                "role": "user" if role != "assistant" else "assistant",
-                "content": [{"type": "input_text", "text": str(content)}],
-            })
-
-        payload: dict[str, Any] = {
-            "model": self._api_model(),
-            "instructions": "\n\n".join(instructions) or "You are a helpful assistant.",
-            "input": input_items,
-            "stream": True,
-            "store": False,
-            "max_output_tokens": max_tokens,
-        }
-        # See ``__init__``. ``summary`` omitted on purpose — it carries
-        # the same TTFT cost on xAI as on Codex and the UI doesn't need it.
-        payload["reasoning"] = {"effort": self._reasoning_effort}
-        if tools:
-            payload["tools"] = self.format_tools(tools)
-        return payload
-
-    def format_tools(self, tools: list[dict[str, Any]]) -> list[dict[str, Any]]:
-        formatted: list[dict[str, Any]] = []
-        for tool in tools:
-            formatted.append({
-                "type": "function",
-                "name": tool.get("name", ""),
-                "description": tool.get("description", ""),
-                "parameters": tool.get("parameters") or tool.get("input_schema") or {},
-            })
-        return formatted
-
-    async def chat(
-        self,
-        messages: list[dict[str, Any]],
-        tools: list[dict[str, Any]] | None = None,
-        max_tokens: int = 8096,
-    ) -> LLMResponse:
-        text_parts: list[str] = []
-        final: LLMResponse | None = None
-        async for chunk, resp in self.stream_chat(
-            messages=messages,
-            tools=tools,
-            max_tokens=max_tokens,
-        ):
-            if resp is not None:
-                final = resp
-            elif chunk:
-                text_parts.append(chunk)
-        return final or LLMResponse(
-            text="".join(text_parts) or None,
-            tool_uses=[],
-            stop_reason="end_turn",
-        )
-
-    async def stream_chat(
-        self,
-        messages: list[dict[str, Any]],
-        tools: list[dict[str, Any]] | None = None,
-        max_tokens: int = 8096,
-        *,
-        abort_signal: Any = None,
-    ) -> AsyncIterator[tuple[str, LLMResponse | None]]:
-        import httpx
-
-        bearer = self._load_bearer_token()
-        payload = self._build_payload(messages, tools, max_tokens)
-        full_content = ""
-        output_items: list[dict[str, Any]] = []
-        input_tokens = 0
-        output_tokens = 0
-        response_status = "in_progress"
-        reasoning_wrapper = _ReasoningStreamWrapper()
-
-        async with httpx.AsyncClient(timeout=self._timeout) as client:
-            async with client.stream(
-                "POST",
-                self._base_url,
-                headers={
-                    "Authorization": f"Bearer {bearer}",
-                    "Content-Type": "application/json",
-                    "Accept": "text/event-stream",
-                },
-                json=payload,
-            ) as resp:
-                try:
-                    resp.raise_for_status()
-                except httpx.HTTPStatusError as exc:
-                    raise RuntimeError(await _http_status_detail("xAI Responses", exc)) from exc
-                event: str | None = None
-                data_lines: list[str] = []
-                try:
-                    async for line in resp.aiter_lines():
-                        if abort_signal is not None and abort_signal.is_set():
-                            break
-                        if line.startswith("event: "):
-                            event = line[7:]
-                        elif line.startswith("data: "):
-                            data_lines.append(line[6:])
-                        elif line == "" and event:
-                            raw = "\n".join(data_lines)
-                            event = None
-                            data_lines = []
-                            try:
-                                data = json.loads(raw)
-                            except json.JSONDecodeError:
-                                continue
-                            event_type = data.get("type", "")
-                            if detail := _responses_sse_error_detail(
-                                "xAI Responses", event_type, data,
-                            ):
-                                raise RuntimeError(detail)
-                            delta = data.get("delta")
-                            # OpenAI Responses SSE emits ``delta`` on multiple
-                            # event types — only ``response.output_text.delta``
-                            # is user-facing text. ``response.function_call_arguments.delta``
-                            # streams the JSON args of an in-flight tool call
-                            # and must NOT leak into the assistant message;
-                            # the canonical args still arrive on
-                            # ``response.output_item.done`` so we can ignore
-                            # the streaming form entirely.
-                            #
-                            # ``reasoning_summary_text.delta`` is no-op here
-                            # in xAI's default config (Grok streams text
-                            # deltas during reasoning anyway) but is wired so
-                            # opt-in ``reasoning_effort`` works symmetrically
-                            # with Codex.
-                            if isinstance(delta, str) and event_type.endswith("reasoning_summary_text.delta"):
-                                yield (reasoning_wrapper.open_chunk(delta), None)
-                            elif (
-                                event_type.endswith("reasoning_summary_text.done")
-                                or event_type.endswith("reasoning_summary_part.done")
-                            ):
-                                # OpenAI Responses emits BOTH events at the
-                                # end of a summary part. ``part.done`` is the
-                                # outer wrapper and is the one we can rely on
-                                # to always arrive — ``text.done`` may be
-                                # omitted by some compat backends. Either
-                                # event closes the ``<think>`` tag so the UI
-                                # surfaces the block before ``output_text``
-                                # starts, which is the whole point of the fix.
-                                closing = reasoning_wrapper.close()
-                                if closing:
-                                    yield (closing, None)
-                            elif isinstance(delta, str) and event_type.endswith("output_text.delta"):
-                                closing = reasoning_wrapper.close()
-                                if closing:
-                                    yield (closing, None)
-                                full_content += delta
-                                yield (delta, None)
-                            item = data.get("item")
-                            if isinstance(item, dict) and event_type.endswith("output_item.done"):
-                                output_items.append(item)
-                            response = data.get("response")
-                            if isinstance(response, dict):
-                                response_status = str(response.get("status") or response_status)
-                                usage = response.get("usage") or {}
-                                if isinstance(usage, dict):
-                                    input_tokens = int(
-                                        usage.get("input_tokens") or usage.get("prompt_tokens") or input_tokens
-                                    )
-                                    output_tokens = int(
-                                        usage.get("output_tokens") or usage.get("completion_tokens") or output_tokens
-                                    )
-                except httpx.HTTPError as exc:
-                    raise RuntimeError(
-                        _stream_interrupt_detail("xAI Responses", exc)
-                    ) from exc
-
-        closing = reasoning_wrapper.close()
-        if closing:
-            yield (closing, None)
-
-        tool_uses: list[ToolUse] = []
-        raw_tool_calls: list[dict[str, Any]] = []
-        for item in output_items:
-            if item.get("type") != "function_call":
-                continue
-            args_raw = item.get("arguments") or "{}"
-            try:
-                args = json.loads(args_raw)
-            except (json.JSONDecodeError, ValueError):
-                args = {"_raw": args_raw}
-            call_id = item.get("call_id") or item.get("id") or str(uuid.uuid4())
-            name = item.get("name") or ""
-            tool_uses.append(ToolUse(id=call_id, name=name, args=args))
-            raw_tool_calls.append({
-                "id": call_id,
-                "type": "function",
-                "function": {"name": name, "arguments": args_raw},
-            })
-
-        stop_reason = "tool_use" if tool_uses else "end_turn"
-        if response_status == "incomplete":
-            stop_reason = "max_tokens"
-
-        raw_message: dict[str, Any] = {"role": "assistant", "content": full_content}
-        if raw_tool_calls:
-            raw_message["tool_calls"] = raw_tool_calls
-        yield ("", LLMResponse(
-            text=full_content or None,
-            tool_uses=tool_uses,
-            stop_reason=stop_reason,
-            input_tokens=input_tokens,
-            output_tokens=output_tokens,
-            raw_message=raw_message,
-        ))
-
-    def format_tool_result(
-        self, tool_use_id: str, content: str, success: bool = True
-    ) -> dict[str, Any]:
-        return {
-            "role": "tool",
-            "tool_call_id": tool_use_id,
-            "content": content if success else f"Error: {content}",
-        }
+from .codex_responses_provider import CodexResponsesProvider
+from .xai_responses_provider import XAIResponsesProvider
 
 
 class LMStudioProvider(_OpenAICompatibleBase):
